moved shutdown_process and get_git_details in core_helper
